Handle parameters of PsiImmediateClassTypes
While seeing which signature files would be updated by a rewritten
toTypeString for kotlin nulls, I noticed there was an incorrect
update for a collection type where the `<E>` parameter dropped
from some method signatures. I found the workaround used for this
in `PsiTypePrinter` and moved it into the psi type initialization.

Bug: 309149849
Test: Added test
Change-Id: I42f6e48462bb4f2d3f0b29e4d5e9aca2681c1db2