Add test demonstrating method comparsion issue with type params
When ComparisonVisitor finds an added method, it double checks that
the signature doesn't match an existing method or inherited method.
Methods match if they have the same name and their parameters match,
which means they have the same type string, same erased type string,
or the old version looks like a type parameter (one uppercase char).

That means if a method has a parameter with a variable type, a new
method with the same name and number of parameters will be matched
against it regardless of the types of the new parameters.

This adds a test demonstrating the issue, and a test where a generic
method should be matched against a child implementation with a
different type.

Bug: 295236822
Test: Added test
Change-Id: If4794c2c19f5200b561d71b4063487b2cb2d02fd