Replace generic parameter types check in method comparison
Instead of calling two types equal if one is generic, this converts
the type from the parent class (performs substitutions of type
parameters) to actually compare them.

This fixes the issue described in the previous CL where an added
method would be compared against an existing method with the same
name as long as the old method had a parameter with a variable type.

Bug: 295236822
Test: CompatibilityCheckTest
Change-Id: Id4e9945506ac4952f8aa4214f162f3dba5707313