Add CodebaseContext to simplify model test suite tests
Previously, model test suite tests either had to use `it` to reference
the `Codebase` being tested, or rename it to something more meaningful
like `codebase`. This change adds a `CodebaseContext` class against
which the test is run to avoid that. This will also allow additional
information to be made available to the tests.

Bug: 317564701
Test: ./gradlew
Change-Id: Ia9f9621a5d86b8023dc91cda17730137479884ef