Hard code value for Float.MIN_NORMAL
In JDK <19 `Float.MIN_NORMAL.toString()` would output `1.17549435E-38f`
but that was incorrect as it is not the shortest string that can
represent that specific floating point number. That was fix in JDK 19
so it now returns `1.1754944E-38f`, unfortunately, that makes the ART
API signature file dependent on the JDK version that produced it.

This change avoids that by hard coding the value. A future change may
switch to using hex floating point for that field (and any other that
uses hex strings in the source) as that is more stable.

Bug: 319819533
Test: ./gradlew
Change-Id: Ie9c8b6a7f8aebd3a0b1090366c2a23a652602d73