Remove filterDuplicates from ModifierList.writeAnnotations
The `filterDuplicates` was added in https://r.android.com/689389 due to
some issue with the UAST that caused some nullability annotations to be
duplicated. The exact bug was not recorded but after some archaelogy
and some additional tests it could not be reproduced. If there is still
an issue then it should be fixed in the Psi model, not in the rest of
the code built on top of it, to ensure that it is consistent with the
other models.

Either way, the existing code is not needed and so this change removes
it.

Bug: 320457204
Test: ./gradlew
Change-Id: Ifadc180f9ee5885dc74884bc9d5ab3c0fb0321c2