Stop returning requiresBody from ModifierListWriter.write()
There was no real benefit in returning `requiresBody` from
`ModifierListWriter.write()`, apart from deduping the logic and it does
not seem a core responsibility of that method. It was originally done
because it looked like it could reuse some of the other logic in the
method, specifically the `requireAbstract` functionality as that was
done in the original code in `JavaStubWriter` but it turns out that
there is no need for it to reuse that.

So, this change extracts methods to check whether a `MethodItem`
requires `abstract` and/or a `body`. It is not clear where the best
place is for the `requiresMethodBodyInStubs(...)` method as it is not
used by `ModifierListWriter` but this will do for now.

Bug: 320457204
Test: ./gradlew
Change-Id: I5495a978978bff6361e94b62822a100ee71d5988