Parameterize testing of TypeItem
Previously, a single test method tested `TypeItem.asClass()` for
multiple different `TypeItem`. That made it difficult to understand
which type items worked and which ones failed. This change creates a
paramaterized test with one test for each `TypeItem`. It replaces the
`check TypeItem asClass()` test and extends it to add tests for a
generic array and an array of primitives which revealed a few more
inconsistencies between the models. It also adds a test for
`hasTypeArguments()` which was not previously tested.

Bug: 313422201
Test: ./gradlew
Change-Id: If1789fbb5b1a7438162319ab6e56273e6ebc938a