Generalize implementation of TypeItem.asClass()
Each of the models had their own top-level (in the model specific
`*TypeItem` base class) implementation of `TypeItem.asClass()` which
handled the subclasses of `TypeItem` differently. This change takes
advantages of the `TypeItem` hierarchy to dedup the common behavior
in default methods in the `TypeItem` interfaces and pushes the model
specific behavior down into the model specific subclasses. There is
still some duplication but that will be cleaned up in future changes
and the behavior is already far more consistent.

Bug: 322465174
Test: ./gradlew
Change-Id: Iabb9686e81dbbfe0f6886f413aef65f52c2a419f