Correct assertion messages
The `assertNotNull(...)` methods were being used incorrectly. The block
is not used to defer the creation of a message but rather is used to
perform some additional checks if the assertion passes. So, the
previous code used the default message if the assertion failed and if
it passed then it created a string and discarded it. This change
corrects that.

Bug: 313422201
Test: ./gradlew
Change-Id: I6ffa1a0b42b7404591332c8001e368689446e013