Make text model suite tests more consistent with source models
The `SourceModelSuiteRunner` will add `getAndroidJar()` to the
classpath to allow tests to access the class. Previously, the
`TextModelSuiteRunner` did not do that. This change fixes that by
creating a `ClassLoaderBasedClassResolver` that will load
placeholder classes from the `getAndroidJar()`.

This change broke a couple of tests for the text model due to
inconsistencies between it and the source models. They will be fixed in
follow up changes.

Bug: 322147308
Test: ./gradlew
Change-Id: I3deb950c1425ed94aeeacea378c0263e867a8dd6