Make TestCodebase.getOrCreateClass(...) require raw class names
Previously, `TextCodebase.getOrCreateClass(...)` would accept class
names with generic arguments, mainly because Metalava did not correctly
handle types. Since the work to implement `TypeItem` correctly that is
no longer necessary so this change removes that capability.

Every call to the method has been checked to see whether it was passing
in a type (which can contain generic arguments) or a raw class name.
The two places where a type was being passed in was for resolving super
classes and interfaces.

They were switched to first create a `TypeItem` (which handles
separating out class names and arguments) and then down cast to
`TextClassTypeItem` to retrieve the `qualifiedName` which was then
passed to `getOrCreateClass(...)`. The down cast is safe because only
class types can be used in `extends` and `implements`.

The names and documentation of `ResolverContext` methods were cleaned
up to make it clear that they are types and not class names.

Bug: 322147308
Test: ./gradlew
Change-Id: I4d7270a859a6d4409241177c0e244f8d307587e5