Correct compatibility check for type parameter in throws list
Previously, it would treat the type parameter as a class whose
qualified name was the name of the type parameter which broke some of
the logic for checking whether two types were equivalent and would not
display useful information about the bounds for the type parameter.
This change adds some extension functions/properties to provide access
to the underlying `throwableClass` and `description`. Eventually, a new
`ThrowableType` will be added that will provide that functionality.

Bug: 322499518
Test: ./gradlew
Change-Id: I2a5dbae2cdbcd43fa1a3e082e6292b6f9700670c