Prune API that is used outside Metalava
There is some code outside Metalava that uses parts of Metalava as a
library. The `@MetalavaApi` annotations are used to annotate those
parts of the API that are used outside. ATM it is just a marker to warn
against removing the API but eventually they may be used to provide a
managed API surface.

Change http://ag/25439473 reduced the use of Metalava APIs outside
Metalava so this change is pruning the API, either by removing
`@MetalavaApi` annotations off the elements that are not used outside,
or removing it altogether as in the case of `arrayDimensions()` and
`bounds()` and the legacy `toErasedTypeString(Item)` which were only
used by that code.

Bug: 303839034
Test: ./gradlew
Change-Id: I5e75e99f61e8da039646742ac8801efc3ef19c2b