Move TextClassItem creation after parsing open brace
Previously, the `TextClassItem` was created after parsing the declared
class type, before parsing the `extends` and `implements` lists.
However, the `extends` and `implements` lists affected the modifiers
and `TextClassItem.classKind` which meant that the latter had to be
mutable. This change moves the creation after the open brace allowing
`TextClassItem.classKind` to be made immutable.

Bug: 322960279
Test: ./gradlew
Change-Id: Ic3d8f0af4ec8b32fb82b5aea7b230493d0386055