Parse typeParameterString into bounds in TextTypeParameterItem.create()
Previously, the `typeParameterString` was a property on
`TextTypeParameterItem` which was parsed on demand into a list of
bounds strings that were then parsed into types. This change pulled
the parsing out into `create(...)` and passed the bounds string list
in instead.

That necessitated regenerating the type parameter string for the
`toString()` method and that revealed a bug in one of the tests that
was missing ` extends java.lang.Object ` which was also fixed.

Some minor improvements were made to the functions that parse the
bounds string list, i.e. renamed and documentation for `bounds()` and
rename and rewrite for `add()` to use standard functions making it much
easier to understand.

Bug: 323270493
Test: ./gradlew
Change-Id: I1ba9bc443e7c90b222c2de7bbfe4a14c5c75c45c