Create super class TypeItems early and remove resolveSuperClasses()
Previously, the `parseClass(...)` method would record the
`superClassTypeString` for later processing into a `TypeItem`. This
change switches it to create `TypeItem` itself and avoid deferring
the work.

It also removes the `ResolverContext` as it is no longer needed to
pass information from the `ApiFile` to the `ReferenceResolver` class.

The changes to the cache stats are interesting. There are 7,849 more
requests. That is due to each class which is spread across multiple
files will resolving its super class, once per file, rather than once
in total. However, there is no increase in the size of the cache so
it is working. The simplification of the code outweighs a few extra
requests.

The number of cache skips (cache requests with a non-empty
`TypeParameterScope` that do not check the cache) has dropped due to
the use of `superObjectType` and `superEnumType`.

Bug: 322458348
Test: ./gradlew
Change-Id: Ib7e040a27c7305338be4f08401889a8127f280ea