Add tests to show super type nullability issue for generic classes
Previous changes tried to fix this issue but did not handle the case
for generic classes. They differ from non-generic classes in that they
provide a non-empty `TypeParameterScope` and so take a different,
uncached, path through the `TextTypeParser.obtainTypeFromString(..)`
method which was not fixed.

This change adds a test to cover that. A follow up change will fix the
issue.

Bug: 323522029
Test: ./gradlew
Change-Id: I238c2abe01e5d1211375a3de71eb1590b904e279