Add BaseTextCodebaseTest.runSignatureTest(...)
Previously, to run signature tests with multiple signature files it was
necessary to wrap them in an `inputSet(...)`. This adds a new function
`runSignatureTest` that takes a varargs and treats them as a single
`InputSet`. It also rewrites all existing uses of `runCodebaseTest`
in the text model tests to use the new function, correcting formatting
where necessary.

Bug: 323614314
Test: ./gradlew
Change-Id: Ie592e5accf229fa16ee7d7211bbde144613664a2